Updated to handle delays elsewhere in the code
I modified the way `countUp` and `countDown` function to subtract off the actual number of ms since the last update, rather than a single `_interval`. This means that the timer still works if `run` is not called as fast as possible. It does mean that the `interval` callback might not fire at precisely the correct time, but it means the timers always run at the correct pace overall (e.g. issue #14).